resume-radar/requirements.txt
pypdf2
pymupdf
python-dotenv
openai
reportlab
pytest   # for running tests